# backend/src/main/java/com/cryptoloan/contract/adapters/in/web/ContractController.java
package com.cryptoloan.contract.adapters.in.web;
import com.cryptoloan.audit.application.AuditService;import com.cryptoloan.contract.application.ContractService;import com.cryptoloan.contract.domain.model.ContractProof;import com.cryptoloan.notification.application.NotificationService;import com.cryptoloan.shared.security.AuthenticatedUser;import org.springframework.http.*;import org.springframework.security.access.prepost.PreAuthorize;import org.springframework.security.core.annotation.AuthenticationPrincipal;import org.springframework.web.bind.annotation.*;
@RestController @RequestMapping("/api/contracts")class ContractController{private final ContractService contracts;private final NotificationService notifications;private final AuditService audit;ContractController(ContractService c,NotificationService n,AuditService a){contracts=c;notifications=n;audit=a;}
 @GetMapping("/{loanId}")ContractProof get(@PathVariable Long loanId,@AuthenticationPrincipal AuthenticatedUser u){ContractProof p=contracts.get(loanId);authorize(p,u);return p;}
 @GetMapping("/{loanId}/document")ResponseEntity<byte[]>document(@PathVariable Long loanId,@AuthenticationPrincipal AuthenticatedUser u){boolean admin=isAdmin(u);byte[] bytes=contracts.document(loanId,u.email(),admin);return ResponseEntity.ok().header(HttpHeaders.CONTENT_DISPOSITION,"attachment; filename=contrat-pret-"+loanId+".txt").contentType(MediaType.TEXT_PLAIN).body(bytes);}
 @PostMapping("/{loanId}/sign/user")ContractProof signUser(@PathVariable Long loanId,@AuthenticationPrincipal AuthenticatedUser u){ContractProof p=contracts.signUser(loanId,u.email());audit.record(u.email(),"CONTRACT_SIGNED_USER","loan:"+loanId,"hash="+p.contractHashHex());notifications.notify(u.email(),"CONTRACT_SIGNED","Contrat signé","Votre signature du contrat #"+loanId+" a été enregistrée.",loanId);return p;}
 @PostMapping("/{loanId}/sign/admin")@PreAuthorize("hasRole('ADMIN')")ContractProof signAdmin(@PathVariable Long loanId,@AuthenticationPrincipal AuthenticatedUser u){ContractProof p=contracts.signAdmin(loanId);audit.record(u.email(),"CONTRACT_SIGNED_ADMIN","loan:"+loanId,"hash="+p.contractHashHex());notifications.notify(p.borrowerEmail(),"CONTRACT_SIGNED_ADMIN","Contrat contresigné","Le contrat #"+loanId+" a été contresigné par CryptoLoan.",loanId);return p;}
 private void authorize(ContractProof p,AuthenticatedUser u){if(!isAdmin(u)&&!p.borrowerEmail().equalsIgnoreCase(u.email()))throw new SecurityException("Contrat d'un autre utilisateur");}private boolean isAdmin(AuthenticatedUser u){return u.authorities().stream().anyMatch(a->a.getAuthority().equals("ROLE_ADMIN"));}}