# backend/src/main/java/com/cryptoloan/contract/domain/model/ContractProof.java
package com.cryptoloan.contract.domain.model;import java.time.Instant;
public record ContractProof(Long loanId,String borrowerEmail,String contractHashHex,String userTxHash,String adminTxHash,String documentReference,Instant generatedAt){public boolean fullySigned(){return userTxHash!=null&&adminTxHash!=null;}public ContractProof userSigned(String tx){if(userTxHash!=null)throw new IllegalStateException("Le contrat est déjà signé par l'utilisateur");return new ContractProof(loanId,borrowerEmail,contractHashHex,tx,adminTxHash,documentReference,generatedAt);}public ContractProof adminSigned(String tx){if(userTxHash==null)throw new IllegalStateException("La signature utilisateur est requise avant la contresignature");if(adminTxHash!=null)throw new IllegalStateException("Le contrat est déjà signé par l'administrateur");return new ContractProof(loanId,borrowerEmail,contractHashHex,userTxHash,tx,documentReference,generatedAt);}}