# backend/src/main/java/com/cryptoloan/identity/adapters/in/web/AuthController.java
package com.cryptoloan.identity.adapters.in.web;
import com.cryptoloan.identity.application.IdentityService;import com.cryptoloan.identity.domain.model.User;import com.cryptoloan.shared.security.AuthenticatedUser;import jakarta.validation.Valid;import jakarta.validation.constraints.*;import java.util.Set;import org.springframework.http.ResponseEntity;import org.springframework.security.core.annotation.AuthenticationPrincipal;import org.springframework.web.bind.annotation.*;
@RestController @RequestMapping("/api/auth") class AuthController{private final IdentityService identity;AuthController(IdentityService identity){this.identity=identity;}@PostMapping("/register")ResponseEntity<Void>register(@Valid @RequestBody RegisterRequest r){identity.register(r.name(),r.email(),r.password());return ResponseEntity.ok().build();}@PostMapping("/login")AuthResponse login(@Valid @RequestBody LoginRequest r){var result=identity.login(r.email(),r.password());return new AuthResponse(result.token(),"Bearer",Profile.from(result.user()));}@GetMapping("/me")Profile me(@AuthenticationPrincipal AuthenticatedUser user){return new Profile(user.displayName(),user.email(),user.authorities().stream().map(Object::toString).collect(java.util.stream.Collectors.toSet()));}record LoginRequest(@Email @NotBlank String email,@NotBlank String password){}record RegisterRequest(@NotBlank @Size(min=2,max=120)String name,@Email @NotBlank String email,@Size(min=8,max=72)String password){}record AuthResponse(String token,String tokenType,Profile user){}record Profile(String name,String email,Set<String> roles){static Profile from(User u){return new Profile(u.name(),u.email(),u.roles());}}}