# backend/src/main/java/com/cryptoloan/notification/domain/model/Notification.java
package com.cryptoloan.notification.domain.model;import java.time.Instant;public record Notification(Long id,String recipient,String type,String subject,String message,Long loanId,Status status,String error,Instant createdAt){public enum Status{PENDING,SENT,FAILED}public static Notification pending(String recipient,String type,String subject,String message,Long loanId){return new Notification(null,recipient,type,subject,message,loanId,Status.PENDING,null,Instant.now());}public Notification sent(){return new Notification(id,recipient,type,subject,message,loanId,Status.SENT,null,createdAt);}public Notification failed(String error){return new Notification(id,recipient,type,subject,message,loanId,Status.FAILED,error,createdAt);}}