# build/site/cryptoloan/1.0/_attachments/openapi.json
{"openapi":"3.0.1","info":{"title":"CryptoLoan API","description":"API REST de gestion des prêts garantis par des crypto-actifs","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Prêts","description":"Gestion des prêts garantis par des crypto-actifs"}],"paths":{"/api/notifications/test":{"post":{"tags":["notification-controller"],"operationId":"test","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Notification"}}}}}}},"/api/loans":{"get":{"tags":["Prêts"],"summary":"Consulter mes prêts","description":"Retourne les prêts de l'utilisateur authentifié.","operationId":"mine_1","responses":{"200":{"description":"Liste des prêts","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LoanResponse"}}}}},"401":{"description":"Authentification requise","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LoanResponse"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Prêts"],"summary":"Créer un prêt","description":"Enregistre une demande de prêt garantie par des crypto-actifs.","operationId":"create","parameters":[{"name":"Idempotency-Key","in":"header","description":"Clé permettant d'éviter la création de doublons","required":false,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Request"}}},"required":true},"responses":{"200":{"description":"Demande enregistrée","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanResponse"}}}},"400":{"description":"Données invalides","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanResponse"}}}},"401":{"description":"Authentification requise","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/contracts/{loanId}/sign/user":{"post":{"tags":["contract-controller"],"operationId":"signUser","parameters":[{"name":"loanId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContractProof"}}}}}}},"/api/contracts/{loanId}/sign/admin":{"post":{"tags":["contract-controller"],"operationId":"signAdmin","parameters":[{"name":"loanId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContractProof"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/loans/{id}/approve":{"patch":{"tags":["Prêts"],"summary":"Approuver un prêt","description":"Permet à un administrateur d'approuver un prêt.","operationId":"approve","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Accès réservé aux administrateurs","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanResponse"}}}},"200":{"description":"Prêt approuvé","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanResponse"}}}},"401":{"description":"Authentification requise","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/notifications":{"get":{"tags":["notification-controller"],"operationId":"mine","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Notification"}}}}}}}},"/api/notifications/stream":{"get":{"tags":["notification-controller"],"operationId":"stream","responses":{"200":{"description":"OK","content":{"text/event-stream":{"schema":{"$ref":"#/components/schemas/SseEmitter"}}}}}}},"/api/loans/all":{"get":{"tags":["Prêts"],"summary":"Consulter tous les prêts","description":"Retourne l'ensemble des prêts. Accès administrateur.","operationId":"all","responses":{"403":{"description":"Accès réservé aux administrateurs","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LoanResponse"}}}}},"200":{"description":"Liste complète des prêts","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LoanResponse"}}}}},"401":{"description":"Authentification requise","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LoanResponse"}}}}}},"security":[{"bearerAuth":[]}]}},"/api/crypto/prices":{"get":{"tags":["price-controller"],"operationId":"many","parameters":[{"name":"cryptoIds","in":"query","required":true,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"number"}}}}}}}},"/api/crypto/price":{"get":{"tags":["price-controller"],"operationId":"one","parameters":[{"name":"cryptoId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"number"}}}}}}},"/api/contracts/{loanId}":{"get":{"tags":["contract-controller"],"operationId":"get","parameters":[{"name":"loanId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContractProof"}}}}}}},"/api/contracts/{loanId}/document":{"get":{"tags":["contract-controller"],"operationId":"document","parameters":[{"name":"loanId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}}}},"/api/auth/me":{"get":{"tags":["auth-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Profile"}}}}}}},"/api/admin/audit":{"get":{"tags":["audit-controller"],"operationId":"recent","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuditEvent"}}}}}}}}},"components":{"schemas":{"Notification":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"recipient":{"type":"string"},"type":{"type":"string"},"subject":{"type":"string"},"message":{"type":"string"},"loanId":{"type":"integer","format":"int64"},"status":{"type":"string","enum":["PENDING","SENT","FAILED"]},"error":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"Request":{"required":["amountEur","collateralAmount","collateralSymbol"],"type":"object","properties":{"amountEur":{"minimum":1.00,"exclusiveMinimum":false,"type":"number"},"collateralSymbol":{"type":"string"},"collateralAmount":{"minimum":0.00000001,"exclusiveMinimum":false,"type":"number"},"liquidationRatio":{"minimum":1.00,"exclusiveMinimum":false,"type":"number"}}},"LoanResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"borrowerEmail":{"type":"string"},"borrowerName":{"type":"string"},"amountEur":{"type":"number"},"collateralSymbol":{"type":"string"},"collateralAmount":{"type":"number"},"liquidationRatio":{"type":"number"},"status":{"type":"string","enum":["PENDING","APPROVED","LIQUIDATED"]},"lastObservedRatio":{"type":"number"},"createdAt":{"type":"string","format":"date-time"},"approvedAt":{"type":"string","format":"date-time"},"liquidatedAt":{"type":"string","format":"date-time"}}},"ContractProof":{"type":"object","properties":{"loanId":{"type":"integer","format":"int64"},"borrowerEmail":{"type":"string"},"contractHashHex":{"type":"string"},"userTxHash":{"type":"string"},"adminTxHash":{"type":"string"},"documentReference":{"type":"string"},"generatedAt":{"type":"string","format":"date-time"}}},"RegisterRequest":{"required":["email","name"],"type":"object","properties":{"name":{"maxLength":120,"minLength":2,"type":"string"},"email":{"type":"string"},"password":{"maxLength":72,"minLength":8,"type":"string"}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"tokenType":{"type":"string"},"user":{"$ref":"#/components/schemas/Profile"}}},"Profile":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"roles":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"SseEmitter":{"type":"object","properties":{"timeout":{"type":"integer","format":"int64"}}},"AuditEvent":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"actor":{"type":"string"},"action":{"type":"string"},"resource":{"type":"string"},"details":{"type":"string"},"occurredAt":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}